+++ include/SI/time.h
@@ -7,7 +7,7 @@
 namespace SI {
 
 template <char _Exponent = 1, class _Ratio = std::ratio<1>,
-          typename _Type = long long int>
+          typename _Type = int64_t>
 struct time_t final : public unit_t<'T', _Exponent, _Ratio, _Type> {
   using unit_t<'T', _Exponent, _Ratio, _Type>::unit_t;
